Refactored SelectorParser logic into LinkedInSelectorParser as first step to creating agnostic selector
