+++ yoga-core/src/main/java/org/skyscreamer/yoga/selector/Selector.java
@@ -3,6 +3,7 @@ package org.skyscreamer.yoga.selector;
 import org.skyscreamer.yoga.populator.FieldPopulator;
 
 import java.beans.PropertyDescriptor;
+import java.util.Map;
 import java.util.Set;
 
 public interface Selector
@@ -15,4 +16,6 @@ public interface Selector
    Selector getField(String fieldName);
 
    Set<String> getFieldNames();
+
+    Map<String, Selector> getFields();
 }
